feat: adicionando status de erro

+++ src/main/java/com/template/business/services/MeteorologiaService.java
@@ -1,10 +1,13 @@
 package com.template.business.services;
 
 import com.template.data.entity.MeteorologiaEntity;
+import com.template.data.exception.MeteorologiaNotFoundException;
 import com.template.data.repository.MeteorologiaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
+
+import java.util.Optional;
 
 @Service
 public class MeteorologiaService {
@@ -17,6 +20,11 @@ public class MeteorologiaService {
     }
 
     public void excluirRegistro(long id) {
-        meteorologiaRepository.deleteById(id);
+        Optional<MeteorologiaEntity> meteorologia = meteorologiaRepository.findById(id);
+        if (meteorologia.isPresent()) {
+            meteorologiaRepository.deleteById(id);
+        } else {
+            throw new MeteorologiaNotFoundException("Para de inventar moda.");
+        }
     }
 }
